# Backend/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Social Media API
  description: A social media API including user, post and comment resources.
  contact:
    name: Asim
    email: adalata@uni.coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://ecology-fiction-3000.codio-box.uk/api/v1'
    description: Development server
  - url: 'localhost:3000/api/v1'
    description: Local server
...
paths:
...
  /users/{id}:
    summary: Represents an individual user.
    description: >
      This resource represents an individual user in the system.
      Each user is identified by a numeric `id`.
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the user.
    get:
      tags:
        - Users
      description:
        Get user by id.
      responses:
        '200':
          description: View of a user's record, excluding password, email, role.
        '404':
          description: User was not found in the DB.
    put:
      tags:
        - Users
      description:
        Update user by ID.
      requestBody:
        description: Existing user data, including edited data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/user
      responses:
        '200':
          description: Update confirmation and link to record.
          content:
            application/json:
              schema:
                $ref: ./post.json#/definitions/postCreated
        '400':
          description: Request data not as expected.
        '401':
          description: Only authenticated users can update a record.
        '403':
          description: Only administrators or the account owner can update the full record.
        '404':
          description: User was not found in the DB.
    delete:
      tags:
        - Users
      description:
        Delete user by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only authenticated users can delete user accounts.
        '403':
          description: Only administrators or account owners can delete user accounts.
        '404':
          description: User was not found in the DB.
  /users:
    summary: Represents the user resource.
    description: >
      This resource represents the system's user resource.
    post:
      tags:
        - Users
      description:
        Create a new user account.
      requestBody:
        description: All new user data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/user
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/userCreated
        '400':
          description: Request data not as expected.
    get:
      tags:
        - Users
      description:
        Retrieve all user accounts.
      responses:
        '200':
          description: All registered users.
        '401':
          description: Only authenticated users can view all users.
        '403':
          description: Only an administrator can view all users.
        '404':
          description: No users found in the DB.
...
  /posts/{postID}/comments/{commentID}/replies/{id}:
    summary: Represents an individual reply.
    description: >
      This resource represents an individual reply in the system.
      The associated post is identified by a numeric `postID`.
      The associated comment is identified by a numeric `commentID`.
      Each reply is identified by a numeric `id`.
    parameters:
        - in: path
          name: postID
          schema:
            type: integer
          required: true
          description: Numeric id of the associated post
        - in: path
          name: commentID
          schema:
            type: integer
          required: true
          description: Numeric id of the comment
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the reply
    get:
      tags:
        - Replies
      description:
        Get reply by id.
      responses:
        '200':
          description: View of a reply.
        '404':
          description: Reply was not found in the DB.
    put:
      tags:
        - Replies
      description:
        Update reply by ID.
      requestBody:
        description: Existing reply data, including edited data
        required: true
        content:
          application/json:
            schema:
              $ref: ./reply.json#/definitions/reply
      responses:
        '200':
          description: Update confirmation and link to record.
          content:
            application/json:
              schema:
                $ref: ./reply.json#/definitions/replyCreated
        '400':
          description: Request data not as expected.
        '401':
          description: Only authenticated users can update replies.
        '403':
          description: Only the owner of a reply or an administrator can update it.
        '404':
          description: Reply was not found in the DB.
    delete:
      tags:
        - Replies
      description:
        Delete reply by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only authenticated users can delete replies.
        '403':
          description: Only the owner of a reply or an administrator can delete it.
        '404':
          description: Reply was not found in the DB.
  /posts/{postID}/comments/{commentID}/replies:
    summary: Represents the reply resource.
    description: >
      This resource represents the system's reply resource.
      The associated post is identified by a numeric `postID`.
      The associated comment is identified by a numeric `commentID`.
    parameters:
        - in: path
          name: postID
          schema:
            type: integer
          required: true
          description: Numeric id of the associated post
        - in: path
          name: commentID
          schema:
            type: integer
          required: true
          description: Numeric id of the comment
    post:
      tags:
        - Replies
      description:
        Create a new reply.
      requestBody:
        description: All new reply data
        required: true
        content:
          application/json:
            schema:
              $ref: ./reply.json#/definitions/reply
      responses:
        '201':
          description: Reply created
          content:
            application/json:
              schema:
                $ref: ./reply.json#/definitions/replyCreated
        '400':
          description: Request data not as expected.
        '404':
          description: Reply was not found in the DB.
    get:
      tags:
        - Replies
      description:
        Retrieve all replies.
      responses:
        '200':
          description: All replies.
        '404':
          description: No replies found in the DB.
...